Moving platform model to model.dart
