+++ lib/src/fitness.dart
@@ -14,7 +14,6 @@ import 'package:path/path.dart' as p;
 import 'code_problem.dart';
 import 'messages.dart' as messages;
 import 'model.dart';
-import 'platform.dart';
 import 'pubspec.dart';
 
 part 'fitness.g.dart';
